% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.GGM.R
\name{Evaluation.GGM}
\alias{Evaluation.GGM}
\title{Evaluation function for the estimated GGM.}
\usage{
Evaluation.GGM(est.precision, true.precision)
}
\arguments{
\item{est.precision}{The estimated precision matrix.}

\item{true.precision}{The true precision matrix.}
}
\value{
A result list including Recall, FDR, F1score, MCC, Hamming Distance,and estimated error of adjacency matrix on F-norm.
}
\description{
Evaluation function for the estimated GGM.
}
\author{
Mingyang Ren \href{mailto:renmingyang17@mails.ucas.ac.cn}{renmingyang17@mails.ucas.ac.cn}.
}
