% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_adj.R
\name{layer_adj}
\alias{layer_adj}
\title{The function of converting the adjacency matrix into the topological layer.}
\usage{
layer_adj(true_adjace)
}
\arguments{
\item{true_adjace}{a p * p adjacency matrix}
}
\value{
Layer_true: a p * 2 matrix to store the information of layer. The first column is the node label, and the second column is the corresponding layer labels.
}
\description{
The function of converting the adjacency matrix into the topological layer.
}
\references{
Zhao, R., He X., and Wang J. (2022). Learning linear non-Gaussian directed acyclic graph with diverging number of nodes. Journal of Machine Learning Research.
}
\author{
Mingyang Ren \href{mailto:renmingyang17@mails.ucas.ac.cn}{renmingyang17@mails.ucas.ac.cn}.
}
